refactor(workflows): share system workflow contract (#1895)
