+++ packages/interfaces/src/automation/index.ts
@@ -2,5 +2,6 @@ export * from './metadata.interface';
 export * from './setting.interface';
 export * from './setting-option.interface';
 export * from './sort.interface';
+export * from './system-workflow-contract.interface';
 export * from './workflow.interface';
 export * from './workflow-builder.interface';
